fix metadata generator (#119)
* fix metadata generator

* fix StaticBuffer constructor

* fix tests and arrange code; but the metadata format still need be fixed

* fix metadata path

+++ examples/erc20/erc20.ts
@@ -180,7 +180,7 @@ export class ERC20 {
             this.storage.balances.set(account, amount);
         } else {
             // @ts-ignore
-            const leftValue = balance + amount;
+            const leftValue = (balance as u128) + amount;
             this.storage.balances.set(account, leftValue);
         }
